Extract synchronous props serialization and parsing logic
