+++ packages/react-native-reanimated/Common/cpp/reanimated/NativeModules/SynchronousPropsBufferSerializer.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <reanimated/Fabric/updates/UpdatesRegistry.h>
+
+#include <vector>
+
+namespace reanimated {
+
+void serializeSynchronousPropsToBuffers(
+    const UpdatesBatch &synchronousUpdatesBatch,
+    std::vector<int> &intBuffer,
+    std::vector<double> &doubleBuffer);
+
+} // namespace reanimated
